Improve documentation for locale import.
When I used the amLocale pipe, I was struggeling with getting the locale loaded. When I did something like

```
{{ myMoment | amLocale:'de' | amDateFormat:'MMMM' }}
```

to thow the month of a date, I always got the english name of the month. After searching the issues, I found this comment https://github.com/urish/ngx-moment/issues/149#issuecomment-315011495, what finally helped me. I checked my node_modules that moment was installed and just did the import. I didn't test in my project, if the manual installation of moment.js is necesssary on a clean project.

I think, it would improve the documentation, if this note is added.